@@ -1,9 +1,19 @@
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:term7moviemobile/libraries/HttpFetch.dart';
+import 'package:get/get.dart';
+import 'package:term7moviemobile/http/HttpFetch.dart';
+import 'package:term7moviemobile/controllers/auth_controller.dart';
 import 'package:term7moviemobile/resourses.dart';
-import 'presenters/router.dart' as router;
+import 'package:term7moviemobile/screens/splash_screen.dart';
+import 'package:term7moviemobile/utils/theme.dart';
+import 'firebase_options.dart';
 
-void main() {
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp(
+    options: DefaultFirebaseOptions.currentPlatform,
+  );
+  Get.put(AuthController());
   runApp(MyApp());
 }
 
@@ -11,12 +21,11 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     config();
-    return MaterialApp(
-      onGenerateRoute: router.generateRoute,
-      initialRoute: '/',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
+    return GetMaterialApp(
+      title: 'F-Cine',
+      theme: MyTheme.myLightTheme,
+      debugShowCheckedModeBanner: false,
+      home: const SplashScreen(),
     );
   }
 
