@@ -1,6 +1,7 @@
 import 'package:flutter/foundation.dart';
 
 class Resourses {
-  static String host =
-      kReleaseMode ? "https://www.fcinema.tk" : "http://10.0.2.2:1024";
+  static String host = "https://www.fcinema.tk";
+  // static String host =
+  //     kReleaseMode ? "https://www.fcinema.tk" : "http://10.0.2.2:1024";
 }
